move pytest_generate_tests to automation conftest.py its very specific to automation running test per genome
